Use LickView in ChatChangeSenderController
